+++ NEWS.md
@@ -1,3 +1,19 @@
-# fcmsafety 1.0.1
+# fcmsafety 0.1.2
 
-* Added a `NEWS.md` file to track changes to the package.
+## Enhancement
+
+1. extract_cid() now support any of the following keys, "InChIKey", "CAS", 
+or "Name", as well as their combinations. 
+2. Add a default value to the cas_col argument in evaluate_compound().
+
+
+## Bug fixes
+
+1. evaluate_compound() "Error: Argument 1 must have names." error fixed.
+2. Remove duplicates results if you have duplicate InChIKey values.
+3. Remove warnings when load_databases().
+
+## Others
+
+1. Retrieval CAS number in the extract_meta() function instead of in evaluate_compound().
+